Make pod-watcher a container service
With this change, instead of having to deploy this code with every
container, it will be deployed once, and watch annotations.

The service will watch for `podwatcher/critical-containers` and
`podwatcher/condition`.

`critical-containers` is a list (comma delimited) of container names
that are the pods required to be running.

If `podwatcher/condition` is empty it will default to `any`, and then
if any critical-container exites, the pod will be deleted. Another
option for `podwatcher/condition` is `all`. When all the containers
have to be complete before they will be deleted.

Once a Pod has been determinted that it needs to be deleted, the
application will attempt to see if an istio contianer is present. If it
is, Istio will be stopped gradefully.

After istio has been stopped, or if there is no istio, the Pod / Job
will be deleted.